🧹 WF8A0X task: split CURATOR harvest modules
