+++ .agentplane/tasks/202607221852-WF8A0X/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607221852-WF8A0X/create-curator-gated-post-task-knowledge-proposa",
   "created_at": "2026-07-30T16:07:19.966Z",
-  "diffstat_sha256": "sha256:c288c59bcb7ae8960e07b4f1a19fa5be07d550f089b1a53995e7ce55fd748691",
+  "diffstat_sha256": "sha256:f7a2f2b25c5c8e65f535b52fec84a61a60113a02bcee4e72e1dacecfa1e574f9",
   "last_verified_at": "2026-07-30T18:43:05.515Z",
   "last_verified_diffstat_sha256": "sha256:b8ce9865041eb716a6b4ed05e91b6c04367a4fda474de8f8cfc86036993fbe26",
   "pr_number": 4694,
